+++ app/controllers/courses_controller.rb
@@ -8,7 +8,7 @@ require_dependency "#{Rails.root}/lib/course_creation_manager"
 require_dependency "#{Rails.root}/app/workers/update_course_worker"
 require_dependency "#{Rails.root}/app/workers/notify_untrained_users_worker"
 require_dependency "#{Rails.root}/app/workers/announce_course_worker"
-require "#{Rails.root}/lib/alerts/deleted_timeline_alert_manager"
+require "#{Rails.root}/lib/alerts/check_timeline_manager"
 
 #= Controller for course functionality
 class CoursesController < ApplicationController
@@ -198,7 +198,7 @@ class CoursesController < ApplicationController
   def delete_all_weeks
     @course = find_course_by_slug(params[:id])
     @course.weeks.destroy_all
-    DeletedTimelineAlertManager.new(@course)
+    CheckTimelineManager.new(@course)
     render plain: '', status: :ok
   end
