+++ app/controllers/weeks_controller.rb
@@ -1,12 +1,12 @@
 # frozen_string_literal: true
-require "#{Rails.root}/lib/alerts/deleted_timeline_alert_manager"
+require "#{Rails.root}/lib/alerts/check_timeline_manager"
 class WeeksController < ApplicationController
   respond_to :json
 
   def destroy
     course = Week.find(params[:id]).course
     Week.find(params[:id]).destroy
-    DeletedTimelineAlertManager.new(course)
+    CheckTimelineManager.new(course)
     render plain: '', status: :ok
   end
 end
