--- lib/alerts/deleted_timeline_alert_manager.rb
@@ -1,16 +1,0 @@
-# frozen_string_literal: true
-
-class DeletedTimelineAlertManager
-  def initialize(course)
-    @course = course
-    create_alerts
-  end
-
-  def create_alerts
-    return unless @course.approved?
-    # it will not create an alert if course doesn't have any training module
-    return if @course.training_module_ids.any?
-    alert = Alert.create(type: 'DeletedTimelineAlert', course_id: @course.id, message: "It appears you have removed either the timeline, week or block from your course.")
-    alert.email_content_expert
-  end
-end
